+++ db/src/main/java/com/psddev/cms/tool/search/ExportContentSearchResultAction.java
@@ -1,10 +1,10 @@
 package com.psddev.cms.tool.search;
 
-import com.psddev.cms.db.ExportContent;
 import com.psddev.cms.tool.Search;
 import com.psddev.cms.tool.SearchResultAction;
 import com.psddev.cms.tool.SearchResultSelection;
 import com.psddev.cms.tool.ToolPageContext;
+import com.psddev.cms.tool.page.ExportContent;
 
 import javax.servlet.ServletException;
 import java.io.IOException;
